Move ingress functions to its own helper. Fix info command
