+++ lib/deploy.js
@@ -20,8 +20,8 @@ const _ = require('lodash');
 const BbPromise = require('bluebird');
 const Api = require('kubernetes-client');
 const helpers = require('./helpers');
+const ingressHelper = require('./ingress');
 const moment = require('moment');
-const url = require('url');
 
 function getFunctionDescription(
     funcName,
@@ -247,97 +247,6 @@ function redeployFunctionAndWait(body, thirdPartyResources, options) {
   });
 }
 
-function getEvents(description) {
-  return !_.isEmpty(description.events) ?
-    description.events :
-    [{ type: 'http', path: '/' }];
-}
-
-function addIngressRuleIfNecessary(functions, options) {
-  const opts = _.defaults({}, options, {
-    verbose: false,
-    log: console.log,
-    namespace: 'default',
-    hostname: null,
-  });
-  const config = helpers.loadKubeConfig();
-  const extensions = new Api.Extensions(helpers.getConnectionOptions(
-    config, { namespace: options.namespace })
-  );
-  const defaultHostname = `${url.parse(helpers.getKubernetesAPIURL(config)).hostname}.nip.io`;
-  const rules = [];
-  _.each(functions, (description) => {
-    const events = getEvents(description);
-    _.each(events, event => {
-      if (event.type === 'http') {
-        const fpath = event.path || '/';
-        if (event.path !== '/' || !_.isEmpty(event.hostname)) {
-          const hostname = event.hostname || opts.hostname || defaultHostname;
-          const absolutePath = _.startsWith(fpath, '/') ?
-            fpath :
-            `/${fpath}`;
-          const previousRule = _.findIndex(rules, r => r.host === hostname);
-          if (previousRule >= 0) {
-            rules[previousRule].http.paths.push({
-              path: absolutePath,
-              backend: { serviceName: description.id, servicePort: 8080 },
-            });
-          } else {
-            rules.push({
-              host: hostname,
-              http: {
-                paths: [{
-                  path: absolutePath,
-                  backend: { serviceName: description.id, servicePort: 8080 },
-                }],
-              },
-            });
-          }
-        }
-      }
-    });
-  });
-  return new BbPromise((resolve, reject) => {
-    if (!_.isEmpty(rules)) {
-      // Found a path to deploy the function
-      const labels = {};
-      _.each(functions, f => {
-        labels[f.id] = '1';
-      });
-      const ingressDef = {
-        kind: 'Ingress',
-        metadata: {
-          name: `ingress-${moment.now()}`,
-          labels,
-          annotations: {
-            'kubernetes.io/ingress.class': 'nginx',
-            'ingress.kubernetes.io/rewrite-target': '/',
-          },
-        },
-        spec: { rules },
-      };
-      extensions.ns.ingress.post({ body: ingressDef }, (err) => {
-        if (err) {
-          reject(
-            'Unable to deploy the ingress rule. ' +
-            `Received: ${err.message}`
-          );
-        } else {
-          if (opts.verbose) {
-            opts.log('Deployed Ingress rule');
-          }
-          resolve();
-        }
-      });
-    } else {
-      if (opts.verbose) {
-        opts.log('Skipping ingress rule generation');
-      }
-      resolve();
-    }
-  });
-}
-
 function deploy(functions, runtime, options) {
   const opts = _.defaults({}, options, {
     hostname: null,
@@ -359,7 +268,9 @@ function deploy(functions, runtime, options) {
           );
         const thirdPartyResources = new Api.ThirdPartyResources(connectionOptions);
         thirdPartyResources.addResource('functions');
-        const events = getEvents(description);
+        const events = !_.isEmpty(description.events) ?
+          description.events :
+          [{ type: 'http', path: '/' }];
         _.each(events, event => {
           const funcs = getFunctionDescription(
               description.id,
@@ -423,7 +334,7 @@ function deploy(functions, runtime, options) {
                       // don't try to add an ingress rule
                     p = new BbPromise((r) => r());
                   } else {
-                    p = addIngressRuleIfNecessary(functions, {
+                    p = ingressHelper.addIngressRuleIfNecessary(functions, {
                       verbose: opts.verbose,
                       log: opts.log,
                       hostname: opts.hostname,
