+++ lib/get-info.js
@@ -20,7 +20,8 @@ const _ = require('lodash');
 const Api = require('kubernetes-client');
 const BbPromise = require('bluebird');
 const chalk = require('chalk');
-const helpers = require('../lib/helpers');
+const helpers = require('./helpers');
+const ingressHelper = require('./ingress');
 
 function toMultipleWords(word) {
   return word.replace(/([A-Z])/, ' $1').replace(/^./, (l) => l.toUpperCase());
@@ -118,12 +119,16 @@ function info(functions, options) {
                 opts.log(`Not found any information about the function "${f}"`);
               } else {
                 const fIngress = _.find(ingressInfo.items, item => (
-                  item.metadata.labels && item.metadata.labels.function === f
+                  item.metadata.labels &&
+                  _.isEqual(item.metadata.labels, ingressHelper.getIngressRuleLabels(functions))
                 ));
                 let url = null;
+                const rule = _.find(fIngress.spec.rules,
+                  r => _.some(r.http.paths, p => p.backend.serviceName === f)
+                );
+                const path = _.find(rule.http.paths, p => p.backend.serviceName === f).path;
                 if (fIngress) {
-                  url = `${fIngress.spec.rules[0].host || 'API_URL'}` +
-                    `${fIngress.spec.rules[0].http.paths[0].path}`;
+                  url = `${rule.host || 'API_URL'}${path}`;
                 }
                 const service = {
                   name: functionService.metadata.name,
