+++ lib/remove.js
@@ -20,48 +20,7 @@ const _ = require('lodash');
 const Api = require('kubernetes-client');
 const BbPromise = require('bluebird');
 const helpers = require('../lib/helpers');
-
-function removeIngressRuleIfNecessary(functions, namespace, options) {
-  const opts = _.defaults({}, options, {
-    verbose: false,
-    log: console.log,
-  });
-  const extensions = new Api.Extensions(helpers.getConnectionOptions(helpers.loadKubeConfig(), {
-    namespace,
-  }));
-  return new BbPromise((resolve, reject) => {
-    extensions.ns.ingress.get((err, ingressInfo) => {
-      const expectedLabels = {};
-      _.each(functions, f => {
-        expectedLabels[f.id] = '1';
-      });
-      const ingressRule = _.find(ingressInfo.items, item => (
-        item.metadata.labels && _.isEqual(item.metadata.labels, expectedLabels)
-      ));
-      if (!_.isEmpty(ingressRule)) {
-        extensions.ns.ingress.delete(ingressRule, (ingErr) => {
-          if (ingErr) {
-            reject(
-              `Unable to remove the ingress rule ${ingressRule}. Received:\n` +
-              `  Code: ${ingErr.code}\n` +
-              `  Message: ${ingErr.message}`
-            );
-          } else {
-            if (opts.verbose) {
-              opts.log(`Removed Ingress rule ${ingressRule.metadata.name}`);
-            }
-            resolve();
-          }
-        });
-      } else {
-        if (opts.verbose) {
-          opts.log('Skipping ingress rule clean up');
-        }
-        resolve();
-      }
-    });
-  });
-}
+const ingressHelper = require('./ingress');
 
 function removeFunction(functions, options) {
   const opts = _.defaults({}, options, {
@@ -97,7 +56,7 @@ function removeFunction(functions, options) {
         } else {
           counter++;
           if (counter === _.keys(functions).length) {
-            removeIngressRuleIfNecessary(
+            ingressHelper.removeIngressRuleIfNecessary(
               functions,
               connectionOptions.namespace,
               { verbose: opts.verbose, log: opts.log }
